// ebk_front/home/church/app/lib/config/interface.ts
export interface Links {
  first: string;
  previous: string;
  next: string;
  last: string;
}
export interface Meta {
  totalItems: number;
  itemCount: number;
  itemsPerPage: number;
  totalPages: number;
  currentPage: number;
}
export interface Token {
  access_token: string;
  refresh_token: string;
}
// end ....... interface generic ....... end

// start ..... interface for get or find data ....... start
export interface Users {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  nom: string;
  prenom: string;
  telephone: string;
  email: string;
  sexe: null | string;
  profil: null | string;
  datenaissance: Date | null;
  adresse: null | string;
  ville: null | string;
  pays: null | string;
  username: null;
  salt: string;
  password: string;
  hashedRt: string;
  privilege: PrivilegesEnum;
  status: string;
  confirm: boolean;
  eglise: Eglise;
}
export interface Eglise {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id_eglise: number;
  nom_eglise: string;
  username_eglise: string;
  photo_eglise: string;
  couverture_eglise: string;
  sigle_eglise: string;
  adresse_eglise: string;
  ville_eglise: string;
  pays_eglise: string;
  nombrefidel_eglise: string;
  status_eglise: string;
  payement_eglise: boolean;
  programme?: Programme[];
  word_of_day_eglise: string;
  about_eglise: string;
  localisation_eglise: string[];
}
// ...
export interface Commentaires {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  commentaire: string;
  users: Users;
}
export interface Favoris {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  users?: Users;
}

export interface Likes {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  users?: Users;
}
export interface Programme {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  titre: string;
  sousProgramme: SousProgramme[];
}
// ...
export interface ItemTesmonial {
  comments: Commentaires[];
  commentsCount: number;
  isLiked: boolean;
  favoritesCount: any;
  isFavorited: any;
  likesCount: any;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  description: null;
  link: string;
  status: string;
  sharecode: string;
  eglise: Eglise;
  favoris: Favoris[];
  commentaire: Commentaires[];
  share: any[];
  likes: Likes[];
  users: Users;
}
// ...
export interface TestmonialsPaginated {
  items: ItemTesmonial[];
  meta: Meta;
  links: Links;
}